refactor(bloc_user): re-import all bloc user in all view
