+++ lib/presentation/pages/login_page.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../bloc/login/login_bloc.dart';
-import '../../bloc/register/register_bloc.dart';
+import '../../bloc/user/login/login_bloc.dart';
+import '../../bloc/user/register/register_bloc.dart';
 import '../../data/localsources/auth_local_storage.dart';
 import '../../data/models/request/login_model.dart';
 import 'home_page.dart';
